move each primitive into its own module; add rough documentation
This commit does not change anything about the implementations.
